requirements.txt
# Core Dependencies
cerebras-cloud-sdk>=1.0.0
opencv-python>=4.8.0
pyttsx3>=2.90
Pillow>=10.0.0
python-dotenv>=1.0.0

# Optional but recommended
numpy>=1.24.0